+++ packages/realm-web-integration-tests/src/remote-mongodb-service.test.ts
@@ -39,10 +39,14 @@ describe("Remote MongoDB", () => {
         await app.logIn(credentials);
     });
 
-    function getCollection() {
-        const mongodb = app.services.mongodb("local-mongodb");
-        const db = mongodb.db("test-database");
-        return db.collection<TestDocument>("test-collection");
+    function getCollection<T extends Realm.Services.MongoDB.Document>() {
+        if (app.currentUser) {
+            const mongodb = app.currentUser.mongoClient("local-mongodb");
+            const db = mongodb.db("test-database");
+            return db.collection<T>("test-collection");
+        } else {
+            throw new Error("Expected an authenticated user");
+        }
     }
 
     let runId: number;
@@ -50,7 +54,7 @@ describe("Remote MongoDB", () => {
         // Genereate a collection
         runId = Math.floor(Math.random() * Number.MAX_SAFE_INTEGER);
         // Insert a couple of documents
-        const collection = getCollection();
+        const collection = getCollection<TestDocument>();
         await collection.insertMany([
             {
                 runId,
@@ -71,7 +75,7 @@ describe("Remote MongoDB", () => {
     });
 
     it("can find documents", async () => {
-        const collection = getCollection();
+        const collection = getCollection<TestDocument>();
         const result = await collection.find(
             { name: { $exists: true } },
             { limit: 10 },
@@ -84,7 +88,7 @@ describe("Remote MongoDB", () => {
     });
 
     it("returns null when finding no document", async () => {
-        const collection = getCollection();
+        const collection = getCollection<TestDocument>();
         const result = await collection.findOne({
             whatever: "non-existent",
         });
@@ -92,7 +96,7 @@ describe("Remote MongoDB", () => {
     });
 
     it("can find a single documents, with projection", async () => {
-        const collection = getCollection();
+        const collection = getCollection<TestDocument>();
         const result = await collection.findOne(
             { runId, name: "Bob" },
             { projection: { name: 1 } },
@@ -108,7 +112,7 @@ describe("Remote MongoDB", () => {
     });
 
     it("can upsert a document if no one is found", async () => {
-        const collection = getCollection();
+        const collection = getCollection<TestDocument>();
         const result = await collection.findOneAndUpdate(
             { runId, name: "Nobody" },
             { name: "Dennis", hiddenField: "a hidden value" },
@@ -125,7 +129,7 @@ describe("Remote MongoDB", () => {
     });
 
     it("can find and update a document, with projection", async () => {
-        const collection = getCollection();
+        const collection = getCollection<TestDocument>();
         const result = await collection.findOneAndUpdate(
             { runId, name: "Bob" },
             { name: "Bobby" },
@@ -142,7 +146,7 @@ describe("Remote MongoDB", () => {
     });
 
     it("can find and replace a document, with projection", async () => {
-        const collection = getCollection();
+        const collection = getCollection<TestDocument>();
         const result = await collection.findOneAndReplace(
             { runId, name: "Alice" },
             { runId, name: "Alison" },
@@ -163,7 +167,7 @@ describe("Remote MongoDB", () => {
     });
 
     it("can find and delete a document", async () => {
-        const collection = getCollection();
+        const collection = getCollection<TestDocument>();
         const countBefore = await collection.count({ runId });
         // Delete the document with the last name (Charlie)
         const result = await collection.findOneAndDelete(
@@ -185,7 +189,7 @@ describe("Remote MongoDB", () => {
     });
 
     it("can aggregate", async () => {
-        const collection = getCollection();
+        const collection = getCollection<TestDocument>();
         const result = await collection.aggregate([
             { $match: { runId } },
             { $group: { _id: null, names: { $push: "$name" } } },
@@ -195,7 +199,7 @@ describe("Remote MongoDB", () => {
     });
 
     it("can count documents, insert a document and count & retrieve it again", async () => {
-        const collection = getCollection();
+        const collection = getCollection<TestDocument>();
         // Determine the number of documents before insertion
         const countBefore = await collection.count();
         // Insert a document
@@ -221,7 +225,7 @@ describe("Remote MongoDB", () => {
     });
 
     it("can insert many documents", async () => {
-        const collection = getCollection();
+        const collection = getCollection<TestDocument>();
         // Determine the number of documents before insertion
         const countBefore = await collection.count();
         // Insert a document
@@ -240,7 +244,7 @@ describe("Remote MongoDB", () => {
     });
 
     it("can delete a document", async () => {
-        const collection = getCollection();
+        const collection = getCollection<TestDocument>();
         const countBefore = await collection.count({ runId });
         // Delete the document with the last name (Charlie)
         const result = await collection.deleteOne({ runId });
@@ -253,7 +257,7 @@ describe("Remote MongoDB", () => {
     });
 
     it("can delete many documents", async () => {
-        const collection = getCollection();
+        const collection = getCollection<TestDocument>();
         const countBefore = await collection.count({ runId });
         expect(countBefore).equals(3);
         // Delete all documents in this run
@@ -267,7 +271,7 @@ describe("Remote MongoDB", () => {
     });
 
     it("can update a document", async () => {
-        const collection = getCollection();
+        const collection = getCollection<TestDocument>();
         // Delete the document with the last name (Charlie)
         const result = await collection.updateOne(
             { runId, name: "Alice" },
@@ -281,7 +285,7 @@ describe("Remote MongoDB", () => {
     });
 
     it("upserts a document when updating and the query match nothing", async () => {
-        const collection = getCollection();
+        const collection = getCollection<TestDocument>();
         // Delete the document with the last name (Charlie)
         const result = await collection.updateOne(
             { runId, name: "Dennis" },
@@ -296,7 +300,7 @@ describe("Remote MongoDB", () => {
     });
 
     it("can update many documents", async () => {
-        const collection = getCollection();
+        const collection = getCollection<TestDocument>();
         // Delete the document with the last name (Charlie)
         const result = await collection.updateMany(
             { runId },
@@ -309,7 +313,7 @@ describe("Remote MongoDB", () => {
     });
 
     it("upserts when updating many non-existing documents", async () => {
-        const collection = getCollection();
+        const collection = getCollection<TestDocument>();
         // Delete the document with the last name (Charlie)
         const result = await collection.updateMany(
             { runId, name: "Dennis" },
@@ -327,10 +331,9 @@ describe("Remote MongoDB", () => {
         this.timeout(10_000);
         this.slow(2_000);
 
-        const mongodb = app.services.mongodb("local-mongodb");
-        const db = mongodb.db("test-database");
-        const collection = db.collection("test-collection");
+        const collection = getCollection<any>();
 
+        // Delete all documents in the collection
         await collection.deleteMany({});
 
         const sleep = async (time: number) =>